# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "type": "object"
          },
          "message": {
            "type": "string"
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": ["code", "message", "requestId"],
        "type": "object"
      },
...
      "CreateOrderDto": {
        "properties": {
          "cardDraftId": {
            "format": "uuid",
            "type": "string"
          },
          "offerId": {
            "format": "uuid",
            "type": "string"
          },
          "quantity": {
            "maximum": 30,
            "minimum": 1,
            "type": "number"
          },
          "recipientAddress": {
            "$ref": "#/components/schemas/PostalAddressDto"
          },
          "selectedAssetId": {
            "format": "uuid",
            "type": "string"
          },
          "senderAddress": {
            "$ref": "#/components/schemas/PostalAddressDto"
          }
        },
        "required": ["cardDraftId", "selectedAssetId", "offerId", "quantity", "recipientAddress", "senderAddress"],
        "type": "object"
      },
...
      "OrderListResponseDto": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/OrderViewDto"
            },
            "type": "array"
          },
          "nextCursor": {
            "format": "uuid",
            "nullable": true,
            "type": "object"
          }
        },
        "required": ["data"],
        "type": "object"
      },
      "OrderResponseDto": {
        "properties": {
          "order": {
            "$ref": "#/components/schemas/OrderViewDto"
          }
        },
        "required": ["order"],
        "type": "object"
      },
      "OrderViewDto": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "currency": {
            "enum": ["CAD"],
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "orderNumber": {
            "type": "string"
          },
          "placedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "recipientAddress": {
            "additionalProperties": true,
            "nullable": true,
            "type": "object"
          },
          "senderAddress": {
            "additionalProperties": true,
            "nullable": true,
            "type": "object"
          },
          "shippingMinor": {
            "minimum": 0,
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "subtotalMinor": {
            "minimum": 0,
            "type": "integer"
          },
          "taxMinor": {
            "minimum": 0,
            "type": "integer"
          },
          "totalMinor": {
            "minimum": 0,
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "orderNumber",
          "status",
          "currency",
          "subtotalMinor",
          "shippingMinor",
          "taxMinor",
          "totalMinor",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "PostalAddressDto": {
        "properties": {
          "city": {
            "maxLength": 120,
            "type": "string"
          },
          "country": {
            "enum": ["CA"],
            "type": "string"
          },
          "line1": {
            "maxLength": 200,
            "type": "string"
          },
          "line2": {
            "maxLength": 200,
            "type": "string"
          },
          "name": {
            "maxLength": 160,
            "type": "string"
          },
          "postalCode": {
            "example": "V6B 1A1",
            "type": "string"
          },
          "region": {
            "maxLength": 3,
            "minLength": 2,
            "type": "string"
          }
        },
        "required": ["name", "line1", "city", "region", "postalCode", "country"],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "Cognito access token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Authenticated Canada-first Souvenote MVP API",
    "title": "Souvenote API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/orders": {
      "get": {
        "operationId": "listOrders",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "type": "number"
            }
          },
          {
            "description": "Opaque owner-scoped cursor from the previous page.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request conflict"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": ["orders"]
      },
      "post": {
        "operationId": "createOrder",
        "parameters": [
          {
            "description": "Unique 16-128 character retry key scoped to the authenticated operation.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 16,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request conflict"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": ["orders"]
      }
    },
    "/api/v1/orders/{orderId}": {
      "get": {
        "operationId": "getOrder",
        "parameters": [
          {
            "in": "path",
            "name": "orderId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponseDto"
                }
              }
            },
            "description": ""
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Authentication required"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Resource not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Request conflict"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Rate limit exceeded"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "",
        "tags": ["orders"]
      }
    },
...
  },
  "servers": [],
...
}